requirements-dev.txt
-r requirements.txt
brainstate

# test requirements
pytest
absl-py
